Doc: forbid JSDoc warnings & work around tag checker
The Popups' JSDocs currently generate without any warnings. Let's keep
it that way for as long as we can by enabling pedantic mode which causes
a failing exit status when warnings are emitted. This behavior can be
verified by adding adding `/** @ignore foo */`.

The JSDoc tag checker should leverage the default enabled dictionaries
but that doesn't seem to be happening[0]. For the time being, allow
unknown tags so that the full range is supported, including @template,
which are very useful for type checking. Minerva already allows unknown
tags.

Lastly, change spaces to tabs since that's what this codebase uses.

[0] https://github.com/jsdoc3/jsdoc/issues/1542

Change-Id: I0aef9f7a6ca4af28d104628cda7763ec70110413