Remove unused attributes from .svg files to optimize them further
An id="…" can potentially be used to target elements individually, manipulate
them, or show them in another context. This is done with the elements in the
file pointer-mask.svg. But not with any other image.

The file-rule="…" only has an effect on shapes that intersect with itself.
This is not the case for this basic "L" shape.

If you think this is fine, I might look at the original OOUI icons as well
and apply the same optimizations on them. But there are 800+ of them to be
looked at, and I want to invest the time only if you think it's worth it.

Change-Id: Ifa4bd60a436e58d3a7c1536f8201fe76fbdac57c